requirements.txt
# 基本機能（ファイル振り分け）
# Tkinterは標準ライブラリに含まれているため、追加のインストールは不要です

# カスタム通知音の生成に必要（オプション）
numpy>=1.20.0

# アイコン生成に必要
Pillow>=9.0.0

# EXE化に必要
pyinstaller>=5.0.0
